chore: rename /users to user and all related files
