+++ packages/server/src/auth/auth.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import * as argon2 from 'argon2';
 import { PrismaService } from '../prisma/prisma.service';
-import { UsersService } from '../users/users.service';
+import { UsersService } from '../user/user.service';
 import { UserSignup, AccessToken } from './types/auth.types';
 
 @Injectable()
