+++ packages/server/src/user/user.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { UsersService } from './users.service';
+import { UsersService } from './user.service';
 import { PrismaService } from '../prisma/prisma.service';
 
 @Module({
